dro/requirements.txt
aif360
matplotlib
numpy
pandas
seaborn
torch
sklearn
jupyter
jupyter_http_over_ws
absl-py
astor
gast
grpcio
h5py
Keras-Applications
Keras-Preprocessing
Markdown
mock
numpy
protobuf
six
tensorboard
tensorflow==1.15.5
tensorflow-estimator
termcolor
Werkzeug
fairlearn==0.4.6
hydra-core